requirements.txt
Flask
Flask_SQLAlchemy
psycopg2-binary
playwright
APScheduler
requests
python-dotenv
twilio
pillow